requirements.txt
backoff==2.2.1
numpy==1.25.1
openai==1.6.1
pandas==2.2.2
python-dotenv==1.0.0
uvicorn==0.27.1
voyageai==0.2.3
fastapi==0.100.0
uvicorn==0.22.0
pandas==2.0.3
pydantic==2.0.3